Customize MSTest runner banner
