+++ src/Adapter/MSTest.TestAdapter/TestingPlatformAdapter/TestApplicationBuilderExtensions.cs
@@ -8,6 +8,7 @@ using Microsoft.Testing.Extensions.VSTestBridge.Capabilities;
 using Microsoft.Testing.Extensions.VSTestBridge.Helpers;
 using Microsoft.Testing.Platform.Builder;
 using Microsoft.Testing.Platform.Capabilities.TestFramework;
+using Microsoft.Testing.Platform.Services;
 
 namespace Microsoft.VisualStudio.TestTools.UnitTesting;
 
@@ -20,7 +21,11 @@ public static class TestApplicationBuilderExtensions
         testApplicationBuilder.AddTestCaseFilterService(extension);
         testApplicationBuilder.AddTestRunParametersService(extension);
         testApplicationBuilder.RegisterTestFramework(
-            _ => new TestFrameworkCapabilities(new VSTestBridgeExtensionBaseCapabilities()),
+            serviceProvider => new TestFrameworkCapabilities(
+                new VSTestBridgeExtensionBaseCapabilities(),
+#pragma warning disable TPEXP // Type is for evaluation purposes only and is subject to change or removal in future updates. Suppress this diagnostic to proceed.
+                new MSTestBannerCapability(serviceProvider.GetRequiredService<IPlatformInformation>())),
+#pragma warning restore TPEXP // Type is for evaluation purposes only and is subject to change or removal in future updates. Suppress this diagnostic to proceed.
             (capabilities, serviceProvider) => new MSTestBridgedTestFramework(extension, getTestAssemblies, serviceProvider, capabilities));
     }
 }
